Log an unrecognized event from MD with all the data
If an event is reported by 'mdadm' which is unknown or ignored
(i.e. hard/impossible to map to a device state change) by the
'md_report' utility, it should log it anyway instead of just
exitting with an error. That way the information is not lost and
can still be extracted from the logs (e.g. using the
SYSLOG_IDENTIFIER field which is the name of the program logging
the message -- "md_report").